:new: Create AKPI manager
Requests made before calling AAA are no longer ignored.
